[Feat] 프론트와의 웹소켓 통신 및 점수 redis 저장
Former-commit-id: 1909eb767bde6de3130f6654fc99caca162c7fdf

+++ backend/spring-server/src/main/java/com/heungbuja/game/state/GameState.java
@@ -23,7 +23,7 @@ public class GameState implements Serializable {
     // ===== 게임 데이터 =====
     private String audioUrl;
     private Map<String, String> videoUrls;
-    private Integer bpm;
+    private Double bpm;
     private Double duration;
     private SectionInfo sectionInfo;
     private SongLyrics lyricsInfo;
